Fix syntax errors in functor implementation
