+++ src/main/scala/application/Menu.scala
@@ -5,12 +5,6 @@ import plants.Plant
 import scala.annotation.tailrec
 import scala.io.StdIn.readLine
 
-import application.FunctorAid._
-import application.FunctorInstances._
-import sensors.SolarSensor
-import plants.SolarPanel
-
-
 // user interface for the entire plant
 case object Menu {
   def start(): Unit = {
@@ -35,7 +29,7 @@ case object Menu {
       case "1" => checkCameras()
       case "2" => checkStorage()
       case "3" => println("IMPLEMENT")        //todo implement
-      case "4" => analyseSolarData() //Functor
+      case "4" => println("IMPLEMENT")
       case "5" => plantTypeMenu()
       case _ => println("Error: invalid input")
 
@@ -75,43 +69,5 @@ case object Menu {
         println("Invalid input")
         checkStorage()
   }
-
- //Analyze solar data working with the functor
-private def analyseSolarData(): Unit = {
-  //informing the user that the fetching and analyzing process started :)
-  println("Fetching and analyzing solar panel data.....")
-
-  //passing into the solar sensor
-  val plant = new SolarPanel(id = 1) //solar panel ID
-  //attach to a plant
-  val sensor = new SolarSensor(plant)
-  //data request to sensor
-  val data = sensor.requestData(sensor.datasetId, sensor.LastMonthTime.toString, sensor.currentTime.toString)
-
-
-  //Functor aid fmap
-  val boostedData = fmap(data) { readings =>
-    readings.map { case (timestamp, value) =>
-      (timestamp, value * 1.1) //energy production boosted by 10%
-    }
-  }
-
   
-  
-  //results and their handling
-  boostedData match {
-    case Right(readings) =>
-      println("\nBoosted Solar Readings:")
-
-      readings.foreach { case (timestamp, value) =>
-        println(s"$timestamp => $value kWh") //val in kWh
-      }
-
-    case Left(error) =>
-      println(s"Error fetching sensor data: $error") //error statement
-  }
-
-  mainMenu() //back to menu function
-}
-
 }
